--- a/.gitignore
+++ b/.gitignore
@@ -261,3 +261,6 @@ examples/Persistence/Persistence/states.db
 
 # MacOS file
 .DS_Store
+
+# Benchmark artifacts
+BenchmarkDotNet.Artifacts/