--- a/src/Proto.Actor/Proto.Actor.csproj
+++ b/src/Proto.Actor/Proto.Actor.csproj
@@ -7,6 +7,7 @@
   </PropertyGroup>
   <PropertyGroup Condition="'$(TargetFramework)' == 'net452'">
     <TargetFrameworkVersion>v4.5.2</TargetFrameworkVersion>
+    <DefineConstants>NET452</DefineConstants>
   </PropertyGroup>
   <PropertyGroup>
     <OutputType>library</OutputType>