--- a/tests/Proto.Actor.Benchmarks/Proto.Actor.Benchmarks.csproj
+++ b/tests/Proto.Actor.Benchmarks/Proto.Actor.Benchmarks.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFrameworks>netcoreapp2.1;netcoreapp3.0;net461</TargetFrameworks>
+    <TargetFrameworks>netcoreapp2.1;net461</TargetFrameworks>
   </PropertyGroup>
 
   <ItemGroup>