Been taking a look at this library for future projects. Saw this issue: https://github.com/AsynkronIT/protoactor-dotnet/issues/432

This PR instantiates `TaskCompletionSource` with the `TaskCreationOptions.RunContinuationsAsynchronously` for targets other than `net452`. In those cases the `WrapTask` hack isn't needed. Also includes suggested fix for the TODO for the timeout handler.

I've included some benchmarks and added a BenchmarkDotNet project to get some more numbers. Results can be reviewed below. These were run on my laptop. Didn't get much difference from the benchmark in the examples folder, but my runs with BenchmarkDotNet shows some difference in both latency and allocations.

#### InProcessBenchmark from examples folder

##### Before change

net452:
Is Server GC False
Dispatcher              Elapsed         Msg/sec
300                     334 ms          48M
400                     306 ms          52M
500                     280 ms          57M
600                     291 ms          55M
700                     296 ms          54M
800                     284 ms          56M
900                     277 ms          58M

netcoreapp2.1:
Is Server GC True
Dispatcher              Elapsed         Msg/sec
300                     264 ms          61M
400                     233 ms          69M
500                     248 ms          65M
600                     256 ms          63M
700                     229 ms          70M
800                     229 ms          70M
900                     238 ms          67M

##### After change

net452:

Is Server GC False
Dispatcher              Elapsed         Msg/sec
300                     332 ms          48M
400                     311 ms          51M
500                     298 ms          54M
600                     325 ms          49M
700                     285 ms          56M
800                     309 ms          52M
900                     284 ms          56M

netcoreapp2.1:

Is Server GC True
Dispatcher              Elapsed         Msg/sec
300                     268 ms          60M
400                     239 ms          67M
500                     232 ms          69M
600                     236 ms          68M
700                     223 ms          72M
800                     229 ms          70M
900                     230 ms          70M

#### New BenchmarkDotNet benchmarks

##### Before change

netcoreapp2.1:

// * Summary *

BenchmarkDotNet=v0.11.5, OS=Windows 10.0.17134.706 (1803/April2018Update/Redstone4)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.0.100-preview4-011223
  [Host] : .NET Core 2.1.9 (CoreCLR 4.6.27414.06, CoreFX 4.6.27415.01), 64bit RyuJIT

Job=InProcess  Toolchain=InProcessEmitToolchain

|            Method |     Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------ |---------:|----------:|----------:|-------:|------:|------:|----------:|
| InProcessPingPong | 5.161 us | 0.1082 us | 0.1867 us | 0.4654 |     - |     - |   1.35 KB |

##### After change

netcoreapp2.1:

// * Summary *

BenchmarkDotNet=v0.11.5, OS=Windows 10.0.17134.706 (1803/April2018Update/Redstone4)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.0.100-preview4-011223
  [Host] : .NET Core 2.1.9 (CoreCLR 4.6.27414.06, CoreFX 4.6.27415.01), 64bit RyuJIT

Job=InProcess  Toolchain=InProcessEmitToolchain

|            Method |     Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------ |---------:|----------:|----------:|-------:|------:|------:|----------:|
| InProcessPingPong | 3.334 us | 0.0665 us | 0.1093 us | 0.3357 |     - |     - |    1000 B |