@PotterDai 2019-05-02
@martinothamar Hi, thank you for the PR! This is amazing actually! It solves a weird bug I encountered when running ClusterStabilityTest in example codes. Previously setResult has certain chances of hanging and I'm really puzzled by that. Thank you very much!
Could you please remove the dotnet 3.0 for now, so that it can pass the checks and I can go ahead and merge it?