@PotterDai 2019-05-02
@martinothamar  Nevermind, I just did the change.
It seems 4.5.2 is causing problems. I'll merge it anyway since I'll start a new PR to strip out dotnet 4.5.2 in this library. = )